test(analysis): create a reproducer test for TC-3286
TC-3286 is about the fact that processing of `q` works differently for
the DB implementation than it does for the in-memory implementation.
This lead to weird behavior in the past. Both implementations must be
aligned to prevent this in the future.

This test gives an example. The test is expected to fail with the
current codebase.

(cherry picked from commit 9368f11eabf1bcccc226509f23a842961e6e6f4f)